rutas correctas

+++ tomainventariosv1.1/src/formularios/primerconteo.java
@@ -483,7 +483,7 @@ public class primerconteo extends JFrame {
 				String codigo= txtcodigo.getText();
 				String marbete= txtbuscar.getText();
 				//String cantidad= txtcantidad.getText();
-				int cantidad= Integer.parseInt(txtcantidad.getText());
+				double cantidad= Double.parseDouble(txtcantidad.getText());
 				String ubicacion= txtubicacion.getText();
 				String almacen= txtalmacen.getText();
 				
@@ -504,7 +504,7 @@ public class primerconteo extends JFrame {
 				} else if(marbete.isEmpty()) {
 					JOptionPane.showMessageDialog(contentPane, "Error ------>>> Verifique el contenido de Marbete");
 				//} else if(String.valueOf(Integer.parseInt(txtcantidad.getText())) != null) {
-				} else if(String.valueOf(Integer.parseInt(txtcantidad.getText())).isEmpty()) {
+				} else if(String.valueOf(Double.parseDouble(txtcantidad.getText())).isEmpty()) {
 					JOptionPane.showMessageDialog(contentPane, "Error ------>>> Verifique el contenido de Cantidad");
 				} else {
 				Connection con = null;
@@ -593,7 +593,7 @@ public class primerconteo extends JFrame {
 				String codigo= txtcodigo.getText();
 				String marbete= txtbuscar.getText();
 				//String cantidad= txtcantidad.getText();
-				int cantidad= Integer.parseInt(txtcantidad.getText());
+				double cantidad= Double.parseDouble(txtcantidad.getText());
 				String ubicacion= txtubicacion.getText();
 				String almacen= txtalmacen.getText();
 				
@@ -612,7 +612,7 @@ public class primerconteo extends JFrame {
 				} else if(marbete.isEmpty()) {
 					JOptionPane.showMessageDialog(contentPane, "Error ------>>> Verifique el contenido de Marbete");
 				//} else if(String.valueOf(Integer.parseInt(txtcantidad.getText())) != null) {
-				} else if(String.valueOf(Integer.parseInt(txtcantidad.getText())).isEmpty()) {
+				} else if(String.valueOf(Double.parseDouble(txtcantidad.getText())).isEmpty()) {
 					JOptionPane.showMessageDialog(contentPane, "Error ------>>> Verifique el contenido de Cantidad");
 				} else {
 				Connection con = null;
